requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
psycopg2-binary>=2.9.0
streamlit
asyncpg
sqlalchemy
